Create Quick Sort ( Efficient )
The time complexity for normal Quick Sort in worst case goes to O(n^2). The worst is when the array is already sorted like 1,2,3,4,5,6,7,8. Because the pivot here will also be the last element.
This can be avoided using Randomized partition.
In this, we don't choose pivot to be the last element, else we choose any random pivot. Complexity almost never goes to O(n^2).